src/FC.CodeFlix.Catalog.Application/UseCases/Genre/ListGenres/ListGenres.cs
using FC.CodeFlix.Catalog.Domain.Repository;

namespace FC.CodeFlix.Catalog.Application.UseCases.Genre.ListGenres;

public class ListGenres : IListGenres
{
    private readonly IGenreRepository _genreRepository;

    public ListGenres(IGenreRepository genreRepository)
        => _genreRepository = genreRepository;

    public async Task<ListGenresOutput> Handle(ListGenresInput request, CancellationToken cancellationToken)
    {
        var searchOutput = await _genreRepository.Search(request.ToSearchInput(), cancellationToken);

        return ListGenresOutput.FromSearchOutput(searchOutput);
    }
}